requirements.txt
customtkinterPyQt6==6.6.1
cryptography==41.0.7
bcrypt==4.1.2
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
pyzmq==25.1.2
pyinstaller==6.3.0
win10toast==0.9
requests==2.31.0
packaging==23.2
google-auth-oauthlib==1.2.0
stripe==6.5.0
paypalrestsdk==1.13.1
